Update middleware.js
Thanks for you tutorial on medium regarding the use of JWT. There was a issue in this code when I was implementing the same (Line : 5). When we use (curl -X GET http://localhost:8000 Note: Header is not provided) command, the token variables is not initialized or undefined. and when we use startsWith Method on undefined it throws an exception of undefined. So I appended a simple (|| "") at the end of the Line. So if the header(['x-access-token'],['authorization']) is not provided the token will be initialized with empty string. Then the method will not throw an error. Again thank you very much for this tutorial.